arrumando bug

+++ Holo/Models/Holograma/CriarHolograma.cs
@@ -3,7 +3,7 @@
     public class CriarHolograma
     {
         public string Descricao { get; set; }
-        public double valor { get; set; }
+        public decimal valor { get; set; }
         public int CategoriaId { get; set; }
         public int Quantidade { get; set; }
     }
